// src/api/openapi/auto-generated/models/MeshType.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An enumeration.
 * @export
 */
export const MeshType = {
    dynamic: 'dynamic',
    fixed: 'fixed',
    unknownDefaultOpenApi: '11184809'
} as const;
export type MeshType = typeof MeshType[keyof typeof MeshType];


export function instanceOfMeshType(value: any): boolean {
    for (const key in MeshType) {
        if (Object.prototype.hasOwnProperty.call(MeshType, key)) {
            if ((MeshType as Record<string, MeshType>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function MeshTypeFromJSON(json: any): MeshType {
    return MeshTypeFromJSONTyped(json, false);
}

export function MeshTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): MeshType {
    return json as MeshType;
}

export function MeshTypeToJSON(value?: MeshType | null): any {
    return value as any;
}